src/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.4.0=pypi_0
ca-certificates=2023.05.30=h06a4308_0
cachetools=5.3.1=pypi_0
certifi=2023.7.22=pypi_0
charset-normalizer=3.2.0=pypi_0
cmake=3.27.2=pypi_0
contourpy=1.1.0=pypi_0
cycler=0.11.0=pypi_0
filelock=3.12.2=pypi_0
fonttools=4.42.0=pypi_0
fsspec=2023.6.0=pypi_0
ftfy=6.1.1=pypi_0
google-auth=2.22.0=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
grpcio=1.57.0=pypi_0
h5py=3.9.0=pypi_0
huggingface-hub=0.16.4=pypi_0
idna=3.4=pypi_0
imageio=2.31.1=pypi_0
importlib-metadata=6.8.0=pypi_0
importlib-resources=6.0.1=pypi_0
jinja2=3.1.2=pypi_0
joblib=1.3.2=pypi_0
kiwisolver=1.4.4=pypi_0
lazy-loader=0.3=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
lit=16.0.6=pypi_0
markdown=3.4.4=pypi_0
markupsafe=2.1.3=pypi_0
matplotlib=3.7.2=pypi_0
mpmath=1.3.0=pypi_0
ncurses=6.4=h6a678d5_0
networkx=3.1=pypi_0
numpy=1.25.2=pypi_0
nvidia-cublas-cu11=11.10.3.66=pypi_0
nvidia-cuda-cupti-cu11=11.7.101=pypi_0
nvidia-cuda-nvrtc-cu11=11.7.99=pypi_0
nvidia-cuda-runtime-cu11=11.7.99=pypi_0
nvidia-cudnn-cu11=8.5.0.96=pypi_0
nvidia-cufft-cu11=10.9.0.58=pypi_0
nvidia-curand-cu11=10.2.10.91=pypi_0
nvidia-cusolver-cu11=11.4.0.1=pypi_0
nvidia-cusparse-cu11=11.7.4.91=pypi_0
nvidia-nccl-cu11=2.14.3=pypi_0
nvidia-nvtx-cu11=11.7.91=pypi_0
oauthlib=3.2.2=pypi_0
opencv-python=4.8.0.76=pypi_0
openssl=1.1.1v=h7f8727e_0
packaging=23.1=pypi_0
pandas=2.0.3=pypi_0
pillow=10.0.0=pypi_0
pip=23.2.1=py39h06a4308_0
plotly=5.16.0=pypi_0
prefetch-generator=1.0.3=pypi_0
protobuf=4.24.0=pypi_0
pyasn1=0.5.0=pypi_0
pyasn1-modules=0.3.0=pypi_0
pyparsing=3.0.9=pypi_0
python=3.9.15=h7a1cb2a_2
python-dateutil=2.8.2=pypi_0
pytz=2023.3=pypi_0
pywavelets=1.4.1=pypi_0
pyyaml=6.0.1=pypi_0
readline=8.2=h5eee18b_0
regex=2023.8.8=pypi_0
requests=2.31.0=pypi_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
safetensors=0.3.2=pypi_0
scikit-image=0.21.0=pypi_0
scikit-learn=1.3.0=pypi_0
scipy=1.11.1=pypi_0
seaborn=0.12.2=pypi_0
setuptools=68.0.0=py39h06a4308_0
six=1.16.0=pypi_0
sqlite=3.41.2=h5eee18b_0
sympy=1.12=pypi_0
tenacity=8.2.3=pypi_0
tensorboard=2.14.0=pypi_0
tensorboard-data-server=0.7.1=pypi_0
threadpoolctl=3.2.0=pypi_0
tifffile=2023.8.12=pypi_0
timm=0.9.5=pypi_0
tk=8.6.12=h1ccaba5_0
torch=2.0.1=pypi_0
torchvision=0.15.2=pypi_0
tqdm=4.66.1=pypi_0
triton=2.0.0=pypi_0
typing-extensions=4.7.1=pypi_0
tzdata=2023.3=pypi_0
urllib3=1.26.16=pypi_0
wcwidth=0.2.6=pypi_0
werkzeug=2.3.7=pypi_0
wheel=0.38.4=py39h06a4308_0
xz=5.4.2=h5eee18b_0
zipp=3.16.2=pypi_0
zlib=1.2.13=h5eee18b_0
